codestyle: Fix import statements ordering and avoid star imports
The CS2103/T style guide[1] specifies that star imports must be avoided
-- imported classes must always be listed explicitly.

The guide also specifies that most fundamental packages must be listed
first. Associated packages should be grouped together with one blank
line between groups.

The code style guide is not clear on what packages are fundamental and
what packages are not, so we follow Eclipse's code style defaults and
order the packages as:

    java.*, javax.*, org.*, com.*, other packages in lexicographical order

Fix violations of these rules in the code base.

There are lots of violations, so the developer here is lazy and simply
used Eclipse's code formatting features on each file. Unfortunately,
eclipse will both reorder imports and expand star imports at the same
time, which is why these seemingly unrelated changes are grouped
together in this commit.