Remove all eclipse project files
Tracking the eclipse project files in the repository causes some
inconveniences for developers.

For instance, when importing the project, eclipse will automatically
modify the project name so that it matches the directory where the
repository was cloned to. For instance, I cloned the repository into a
directory called "CS2103AUG2016-T11-C4-main", and so when I import the
project Eclipse automatically makes the following changes:

    diff --git a/.project b/.project
    index 1c9339c..03116ac 100644
    --- a/.project
    +++ b/.project
    @@ -1,7 +1,7 @@
     <?xml version="1.0" encoding="UTF-8"?>
     <projectDescription>
    -       <name>addressbook-level4</name>
    -       <comment>Project addressbook-level4 created by Buildship.</comment>
    +       <name>CS2103AUG2016-T11-C4-main</name>
    +       <comment>Project CS2103AUG2016-T11-C4-main created by Buildship.</comment>
            <projects>
            </projects>
            <buildSpec>

This is annoying because it means I have to make sure I do not
accidentally commit such changes.

Furthermore, Eclipse also stores some developer-specific information in
its project files as well. For instance, to improve my code quality I
use the checkstyle[1] plugin in Eclipse. However, enabling this plugin
adds the following lines:

    diff --git a/.project b/.project
    index 1c9339c..22caba0 100644
    --- a/.project
    +++ b/.project
    @@ -15,9 +15,15 @@
                            <arguments>
                            </arguments>
                    </buildCommand>
    +               <buildCommand>
    +                       <name>net.sf.eclipsecs.core.CheckstyleBuilder</name>
    +                       <arguments>
    +                       </arguments>
    +               </buildCommand>
            </buildSpec>
            <natures>
                    <nature>org.eclipse.buildship.core.gradleprojectnature</nature>
                    <nature>org.eclipse.jdt.core.javanature</nature>
    +               <nature>net.sf.eclipsecs.core.CheckstyleNature</nature>
            </natures>
     </projectDescription>

Which is again annoying because I need to make sure that I do not commit
this change. However, I can't just revert this change, as I need this
change to remain in the .project file to keep the checkstyle plugin
enabled for this project. At the same time, I do not want to impose my
usage of the checkstyle plugin on other developers.

The Eclipse project files seem to be a very developer-specific. As such,
to save developers the trouble, let's just remove all Eclipse project
files from the repository. After all, we can always regenerate them with
the "Gradle->Gradle Project" importer.

Also, modify the .gitignore to ignore any generated eclipse project
files.

[1] http://eclipse-cs.sourceforge.net/