build.gradle: remove project.ext block
The project.ext block contains all the version numbers for the
dependencies. Other than that, the project.ext properties are not used
at all.

We can directly specify these dependency versions in the 'dependencies'
block, saving lots of lines in the build.gradle file and making the file
look cleaner and neater.