ui: Simplify UiPart
The ui subsystem tries to encapsulate the different logical parts of a
user interface through a "UiPart" system. However, this system is
needlessly complicated.

Let's take a look, for example, at how a UiPart is initialized. We shall
use CommandBox as an example:

1. First, we need to create a "placeholder" AnchorPane in our FXML file.

2. Then, we call CommandBox.load(primaryStage, commandBoxPlaceholder etc.

3. CommandBox.load() constructs a "sample CommandBox" object to pass to
   UiPartLoader.loadUiPart()

4. CommandBox.load() then calls UiPartLoader.loadUiPart() with the
   primaryStage, commandBoxPlaceholder and sampleUiPart.

5. UiPartLoader.loadUiPart() constructs a FXMLLoader instance

6. UiPartLoader.loadUiPart() calls getFxmlPath() on the sampleUiPart to
   get the FXML file path

7. UiPartLoader.loadUiPart() then calls fxmlLoader.load() to load the
   FXML view

8. fxmlLoader.load() will load the JavaFX UI nodes in the FXML file, and
   constructs YET another new instance of CommandBox, called `controller`.

9. UiPartLoader.loadUiPart() calls controller.setNode(),
   controller.setPlaceholder(), etc. to configure it.

10. UiPartLoader.loadUiPart() returns the new CommandBox instance

11. CommandBox.load() calls configure() on the new CommandBox instance,
    to further configure it.

As we can see, CommandBox and UiPartLoader are extremely coupled with each
other -- there is a lot of back-and-forth function calling between them.

Furthermore, due to the back-and-forth-ing, implementing a new UiPart is unnecessarily hard. We need to implement:

* setNode()

* getFxmlPath()

* setPlaceholder()

* configure()

and think of what is the sane thing to do in each of these functions.

Generally speaking, the whole UiPart system is a huge mess. The main
reason why it is such a huge mess is because the parent-child
relationship between a controller class and its associated fxml view is
not so clear. FXML allows developers to specify a "fx:controller"
attribute, and FXMLLoader will construct a new instance of that
controller class when loading the FXML file. This is the approach which
the current code uses. In this system, the controller class is a child
of the view.

However, the major drawback of such an approach is that we cannot pass
arguments to the constructor of the controller class. As such, this code
attempts to work around it by having the whole UiPartLoader and
configure() dance, but it ends up being a huge mess as we can see above.

If we think about it, the major reason why we have so much trouble with
the above approach is because letting the view initiate the controller
is the wrong way to go about it.

So, let's ditch the whole "fx:controller" thing. Instead, rather than
letting the FXML Loader create the instance of our controller class,
*WE* create the controller class, and then pass it to the FXML Loader.
The controller class is now the parent, and the FXML view the child.
This places how we construct the object directly in our control, which
means we can have much richer constructors that are fully representative
of what data the controller classes encapsulate, without needing to
resort to separate setNode(), getFxmlPath(), setPlaceholder(),
configure() methods to fully configure the controller instance.

With this commit, this is what the constructor of `UiPart` will now do.
And now, we can fully ditch UiPartLoader, and in the process remove lots
of unnecessary code. It is also much easier to create a UiPart now, as
there are no more abstract methods that need to be implemented. A simple
UiPart subclass can look like:

    public class NewUiPart extends UiPart<VBox> {
        private final Object arg;

        @FXML
        private TextBox myNode;

        public NewUiPart(Stage primaryStage, Object arg) {
            super("/view/NewUiPart.fxml", primaryStage);
            // We can initialize final instance members!
            this.arg = arg;
            // We can even start accessing nodes in our FXML file
            // immediately!
            myNode.setText("hi");
        }
    }

One problem is that without the "fx:controller" attribute, JavaFX FXML
editors will not be able to link the controller class with the FXML
file, as they do not know what the controller class is. However, given
the amount of code we remove in this commit, this seems to be a
worthwhile tradeoff.