storage: implement JsonTaskBookStorage and related classes
We currently use JAXB for serializing/deserializing the TaskBook
database as an XML file. At the same time, we use Jackson to
serialize/deserialize the Config and UserPrefs.

To reduce the cognitive overhead of having to understand two separate
libraries for serialization/deserialization, we should pick only one --
either JAXB (XML) or Jackson (JSON).

In this case, we are going to go with Jackson, as JSON is a much simpler
format compared to XML. Also, as shown in this commit, Jackson has the
ability to "mixin" annotations into the model classes, which allows us
to properly serialize/deserialize our model classes without having to
resort to adaptor classes (e.g. XmlAdaptedTask,
XmlSerializableTaskBook). This allows us to implement proper
serialization/deserialization with very little code.

This commit focuses on implementing JsonTaskBookStorage and its
associated classes to fully implement JSON task book
serialization/deserialization. In later commits we will switch the code
base to use only JsonTaskBookStorage, so we can remove the xml task book
storage functionality.