model: move TaskNotFoundException to a new file
TaskNotFoundException is currently a nested class within UniqueTaskList.

TaskNotFoundException is useful when we want to thrown an exception
signifying that a requested task does not exist.

However, the UniqueTaskList class will soon be deleted when we remove
the unique task constraint. So, move the TaskNotFoundException out from
within UniqueTaskList to a new file.