model: remove unique tasks constraint
The current implementation of UniqueTaskList is not so useful to the
user.  It only checks to see if no two tasks in the task list are
exactly the same. (e.g. if a task name differs by just one letter, they
are considered different).

As can be seen from the diffstat of this commit, lots of code goes into
implementing this unique task constraint. On the other hand, in practice
it does not bring much value to the user. Our underlying list data
structure is also perfectly capable of handling duplicate tasks.

So, let's remove the unique task constraint and delete lots of code with
it.