ReadOnlyTaskBook/TaskBook: merge getTasks() and getTaskList()
ObservableList is a subtype of List. So, by pulling up the getTasks()
interface into ReadOnlyTaskBook will not make that much of a difference:
callers not interested in the return type being an ObservableList can
just simply cast it to List and be on their way.

At the same time, we get to make the code simpler.