Model: make task list filtering interface open to extension
Currently, the updateFilteredTaskList() method of the Model interface
has the following overloads:

* If updateFilteredTaskList() is called with no arguments, then all task
  items in the list will be shown.

* If updateFilteredTaskList(keywords) is called with a set of keywords,
  then the tasks in the list will be filtered based on whether any words
  in their name matches one of the keywords.

This is not very good API design as it is not open to extension. In the
future, we may wish to add more different filtering methods. Such as,
e.g. filtering based on whether tasks are done or not. However, with the
"overload a single method" design that updateFilteredTaskList() has,
there is simply no way we can extend this method with new filtering
methods, short of introducing a new type.

So, let's redesign this API. Standing back a little, we can see that the
filtering functionality just needs a Predicate which would tell, for
each task, whether the task should be shown in the list or not.

So let's just expose the filtering functionality fully with a
`setFilter()` method. Callers can simply pass their own Predicate in to
filter the list to their liking, making the list filtering
functionality open to extension.