UiManager: keep own reference to primaryStage
Currently, all UiParts store a reference to their primaryStage, which
means that they can modify the primaryStage at their own whim and fancy.

We want to prevent that -- UiParts should not be able to access the
primaryStage as it makes it hard to reason about which UiPart is
controlling the primaryStage and which is not. As such, we want to
remove the reference to the primaryStage from all UiParts.

Before we make this change, however, we note that UiManager relies on
MainWindow (a UiPart) to keep its reference to the primaryStage. Since
UiParts should not be storing a reference to the primaryStage, we move
the reponsibility to the UiManager itself.