UiPart: push down primaryStage
Currently, all UiParts store a reference to their primaryStage, which
means that they can modify the primaryStage at their own whim and fancy.

We want to prevent that -- UiParts should not be able to access the
primaryStage as it makes it hard to reason about which UiPart is
controlling the primaryStage and which is not. As such, we want to
remove the reference to the primaryStage from all UiParts.

Of all the UiParts, we note that only MainWindow and HelpWindow require
a reference to the primaryStage. All other UiParts do not, and SHOULD
NOT, keep a reference to the primaryStage.

So, we push down the `primaryStage` field from UiPart to MainWindow and
HelpWindow only. All other UiParts thus effectively lose access to the
primaryStage, which is a good thing for encapsulation.