JsonTaskBookStorage: register JavaTimeModule in ObjectMapper
We are going to add java.time objects to our data structures, and as
such jackson need to know how to serialize/deserialize them.

By default, Jackson does not know how to serialize/deserialize java.time
objects. However, the jackson-datatype-jsr310 package's JavaTimeModule
is able to teach Jackson how to serialize/deserialize java.time objects.

As such, we register the JavaTimeModule with the ObjectMapper that is
created by default by JsonTaskBookStorage.