ui: restore status bar footer
The status bar footer was mistakenly removed in 83d8b4f (ui: Simplify
UiPart, 2016-10-03). This commit restores it.

However, it turns out that StatusBarFooter was also actually broken
by the aforementioned commit, as its `mainPane` and `placeHolder` fields
are now both null, so a NullPointerException will be thrown when the
StatusBarFooter is initialised. Fix this as well.