UiManager.showAlertDialogAndWait(): only set owner if it has a scene
In UiManager::start(), an exception may occur in the constructor of
MainWindow. This exception is thrown before primaryStage.setScene() is
called, resulting in primaryStage having a null Scene.

The try-catch handler then calls showFatalErrorDialogAndShutdown(),
which in turn calls showAlertDialogAndWait().

showAlertDialogAndWait() effectively calls
alert.initOwner(primaryStage). However, alert.initOwner() requires the
Stage to have a non-null scene. At this point, as mentioned above,
primaryStage has a null Scene, so as a result a NullPointerException is
thrown, resulting in no alert dialog being shown.

Fix this by not calling alert.initOwner() if the owner is null or its
scene is null. After all, it is not so important that the alert dialog
has an owner.