ui: replace AnchorPanes with UiRegions
From the FXML files, it is not so obvious that the AnchorPanes are meant
to be filled in with UiParts.

To make it more obvious, we introduce a new control called `UiRegion`.
This control only has two methods: setNode() and getNode(), which can be
used to fill the region up with the root node of the UiPart.

As can be seen in the diff, this simplifies a lot of code and makes it
easier to read.

It should be noted that the padding information has been removed for now
to keep the FXML file simple. This is a style property that should be
implemented in the stylesheet instead. This will be done in the future.