build.gradle: introduce 'pandoc' task
This task will generate a PDF file in `build/docs/manual.pdf` containing
a combination of `docs/UserGuide.md` and `docs/DeveloperGuide.md`, in
that order.

Developers must run it explicitly with `gradle pandoc`. It is not run by
default because it requires that developers properly set up pandoc and
xelatex with the appropriate packages in the system.