DeveloperGuide: add a nice LaTeX glossary
In the PDF output, the "Glossary" appendix is totally blank.

This is because the pandoc LateX generator does not understand HTML, and
thus ignores the glossary list.

To make our glossary actually visible, let's add a custom hand-crafted
LaTeX glossary. Yes, we have to repeat glossary both times, but we
shouldn't have that many glossary items.