MainApp: make fields private
In the past, there existed a class called TestApp which would inherit
from MainApp. This TestApp would override the protected methods and
access the protected fields of MainApp in order to stub in its own logic
for testing purposes.

However, we don't have TestApp now, and in the future we are going to
implement tests that do not require stubbing out MainApp methods. So,
let's make MainApp's fields and methods truly private.