MainApp::stop(): don't terminate process
We are going to add a MainApp integration test in the next commit, which
will initiate MainApp and run it as a JavaFX Application.

However, MainApp's stop() method calls System.exit(). This is bad
because tests are run in the same process, but System.exit() will cause
the test process to terminate prematurely, which will result in not all
tests being run.

Fix this by removing the call to System.exit(). At the same time, remove
Platform.exit() as well -- the JavaFX docs[1] state that when stop() is
called, the application is already exiting anyway.

[1] https://docs.oracle.com/javase/8/javafx/api/javafx/application/Application.html