model: remove UserPrefs and GuiSettings
It was being very annoying and getting in the way of the Config rework.
It only stores the GuiSettings, but yet a lot of code is needed to
support it.

For now, to make it easier to morph the code, let's just remove it. The
GuiSettings can be easily brought back once the Config rework is
finished.