storage: implement JsonConfigStorage
We are going to make Config a part of Model. According to the overall
architecture design, the Model has a corresponding Storage for
saving/loading the model. For example, TaskBook (a model class) has a
corresponding TaskBookStorage.

So likewise, for the Config class, we need to have a corresponding
ConfigStorage. The Config format is JSON, so we call the implementing
class JsonConfigStorage.