ModelManager: manage Config as well
We are going to implement a `setdatadir` command which will change the
taskbook storage file location.

To be able to implement such a command, commands need to have access to,
and be able to modify, the Config.

At the same time, we need the Model to be aware of any changes to the
Config, for e.g. undo/redo support. As such, we need to move the Config
into the Model.