Move "save-taskbook-when-changed" handling from Storage to Logic
Currently, Storage will automatically save the task book when the
TaskBookChangedEvent is posted. This is extremely bad, because it breaks
encapsulation.

For example, imagine that we have two task books, A and B. We also have
two storage instances for each. Now, with the current implementation if
the TaskBookChangedEvent is fired for task book A, BOTH storage
instances will write the contents of task book A to their storage file!
Effectively, task book B's contents would be overwritten!

To fix this broken design, we instead make this auto-saving of task
books explicit by moving it into the Logic component. In the logic
component, we are able to check if commands have modified the task book,
and then save the task book if they have. This fixes the problem
described above.