ObservableListChangeListener: keep references to the ObservableLists
There is a bug where even though the task book has changed, the task
book was not being saved at all.

Presently (in LogicManager.java), we use a TaskBookChangeListener (which
inherits from ObservableListChangeListener) to monitor a task book from
changes. However, seemingly despite the fact that the task book was
modified, the ObservableListChangeListener was still returning
hasChanged = false!

It turns out that because ObservableListChangeListener did not keep a
reference to the ObservableLists it was supposed to observe, these
ObservableLists were getting GC'd. As such, the
ObservableListChangeListener did not get the change events that it
expected.

Fix this by keeping a reference to the observable lists.