parser: implement IndexParser
This allows us to parse task indexes with the CommandLineParser.

Take note that the `IndexParser` will convert 1-indexes to 0-indexes in
the parser itself. I think that this is much better as it cleanly
separates the user interface (task indexes start at 1), with our
underlying data model (task indexes start at 0).