Model: undo/redo: remove dependency on the logic component
The undo/redo methods in Model take a `Command`, which belongs to the
logic component. This means that the model component now depends on the
logic component, which is bad because the logic component also depends
on the model component (aka it makes the dependency bi-directional).

It also makes the undo/redo system hard to test because we must create
dummy Commands just to use it (see the test cases).

But, why does the undo/redo system need the `Command`? Well, it is just
so it can give the commit a name. Using a String for the name
will do the exact same thing, so let's do that.