Implement sticky indexes (implement sorting of tasks)
The main problem currently with listing tasks in our user interface is
that we do not have any auto-sorting of tasks. However, let's say we
have two deadline tasks, which we want to auto-sort in ascending due
datetime (earliest first):

d1. "A" Due 10am
d2. "B" Due 11am

Now, let's say we modify "B" so that it is due at 9am. The listing can
be either like this:

d1. "B" Due 9am
d2. "A" Due 10am

Or it can be like this:

d2. "B" Due 9am
d1. "A" Due 10am

This commit implements the latter, because keeping the task indexes
"stuck" to the task, in my opinion, is less likely to confuse the user
when the task changes position in the list.

The way we go about doing this is to introduce a new interface
`IndexedItem<T>`, which attaches a "working index" to an item. This
"working index" will be stuck to the item until the list gets
repopulated.

And now, instead of the Model exposing floating/deadline/event task
lists of ObservableList<Optional<T>>, it now exposes
ObservableList<IndexedItem<T>>.

In other news, the nitty gritty of implementing the "working indexes"
has been moved into a separate class called "WorkingTaskBook".