Model: undo/redo: store config as well
Currently, if I set the taskbook data storage location with

    setdatadir b

I can't undo it, because the undo/redo system does not know of the
configuration.

So let's teach the undo/redo system how to store the configuration
state, as well as how to restore it, and detect if it has been modified.