JsonTaskBookStorage::moveTaskBook(): only move when source file exists
When the current storage file does not exist (which can happen when the
application is run for the first time), `moveTaskBook()` will throw an
error as it attempts to move an non-existent file.

So, let's teach `moveTaskBook()` to not move the file if the source file
does not exist in the first place.