model: keep track of selected task
Currently, when we add/edit a task, the UI does not highlight (select)
the added/edited task. This is because the Model layer does not have any
concept of a "selected" task.

So, let's teach the model what a "selected" task is. At the same time,
we modify all the add/edit/delete methods to select the
added/edited/deleted task appropriately (For deletion, we clear the
select if the deleted task was selected).

Furthermore, we can also teach the UI to select the selected task on the
UI side, so the user will get visual feedback on what task is selected.