Hotfix for 'filter-rework'
Turns out that when hiding and showing the MainWindow (e.g. using the
Ctrl-Space feature), all the fontawesome icons become squares. This is
because JavaFX does some funky stylesheet stuff, and the
"-fx-font-family: Open Sans" ends up overriding the glyph-icon font. We
work around this by explicitly setting the font family of glyph-icon to
FontAwesome.