Remove EventBus
Currently, the event bus is only used for showing the "Task book last
modified" status in the UI. Yes, the entire event system is used just
for one status message, which is itself not really that useful and, if
really needed, could be re-implemented with much less code.

But that's not the real problem though -- The real problem is that the
even bus system is bad for encapsulation. The "EventsCenter" is a global
singleton, and all listeners for events of type T will receive all of
its events, no matter which object actually sent it. This means that it
is impossible to cleanly separate an instance of the task tracker
application from another.

An example: Let's say we implement a feature that allows the user to run
multiple instances of the application in the same process. We will thus
have two Models, A & B. We will also have two status bars (Task book
last modified status), sA and sB.  However, if only A gets modified,
BOTH sA and sB will update their status, which is totally wrong since
model B was not modified at all.

Furthermore, relying on events means that we need to pepper the code
with event raising functions in order to satisfy what listeners expect.
However, for complex systems, this is a potential source of bugs. With
the current code of Task Tracker, for example, if you try to run "clear
fin", and finished tasks were actually cleared, you will notice the
"Task book last modified" status still does not change at all. This is
because the developer of the code forgot to add in the
"indicateTaskBookChanged()" method call to the relevent code paths.

To conclude, the event bus enforces poor encapsulation and is an extra
source of bugs, so let's remove it.